% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_all_gradient.R
\name{all_gradient_CAU}
\alias{all_gradient_CAU}
\title{Scoring Method: CAU (0-95th)}
\usage{
all_gradient_CAU(metrics.df, sensitivity.df, first.metric)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}
}
\value{
*** NOT COMPLETE ***. Scores the samples.
}
\description{
Scoring Method: CAU (0-95th)
}
