% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_all_gradient.R
\name{all_gradient_min_95}
\alias{all_gradient_min_95}
\title{Scoring Method: All Site Classes Gradient (Min - 95th)}
\usage{
all_gradient_min_95(metrics.df, sensitivity.df, first.metric)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}
}
\value{
Scores the samples.
}
\description{
Scoring Method: All Site Classes Gradient (Min - 95th)
}
