% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_scoring.R
\name{bal_score_methods}
\alias{bal_score_methods}
\title{balanced score methods}
\usage{
bal_score_methods(merged.df, method)
}
\arguments{
\item{merged.df}{=}

\item{method}{= two method options exist. The first option is "UNEVEN", which
uses the reference and degraded distributions 5th and 95th percentiles to
define scoring thresholds.  "EVEN" uses the median of the reference
distribution and...}
}
\value{
Balanced scoring...
}
\description{
balanced score methods
}
