% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_ce.R
\name{balanced_ce}
\alias{balanced_ce}
\title{Balanced Classification Efficiency}
\usage{
balanced_ce(metrics.df, first.metric, quant.df, condition.colname, ref.cond,
  deg.cond, ref.df, deg.df, quant.ref)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{quant.df}{= Data frame containing ref.cond quantile values.}

\item{ref.cond}{= the site class that represents the better condition.}

\item{deg.cond}{= the site class that represents the poorer condition.}

\item{ref.df}{= a data frame of only the ref.cond values.}

\item{quant.ref}{= a data frame of reference quantile values.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Balanced Classification Efficiency
}
