% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_scoring.R
\name{balanced_scoring}
\alias{balanced_scoring}
\title{Balanced Scoring}
\usage{
balanced_scoring(metrics.df, sensitivity.df, first.metric, method)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{sensitivity.df}{= the data frame output from the metric_sensitivity
function where method was set to "BDE" or "ALL."}

\item{first.metric}{= the first metric column that appears in your wide
data frame when reading from left to right. It is assumed that all columns
to the right of the first metric column are metrics. Any other columns will
result in an error or, if the column is numeric, it will be treated as if
it were a metric.}

\item{method}{= two method options exist. The first option is "UNEVEN", which
uses the reference and degraded distributions 5th and 95th percentiles to
define scoring thresholds.  "EVEN" uses the median of the reference
distribution and...}
}
\value{
Balanced scoring...
}
\description{
Balanced Scoring
}
