% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{barbour}
\alias{barbour}
\title{Barbour et. al 1996}
\usage{
barbour(prep.sensitivity, .metric_col)
}
\arguments{
\item{prep.sensitivity}{PARAM_DESCRIPTION}

\item{.metric_col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
}
