% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barbour.R
\name{barbour_sensitivity}
\alias{barbour_sensitivity}
\title{Barbour Sensitivity Score}
\usage{
barbour_sensitivity(barbour.df, score)
}
\arguments{
\item{barbour.df}{= data frame of metrics scored using the Barbour et al. 1996 method
for testing metric sensitivity}

\item{score}{= a number 0 - 3 indicating which sensitivity score will be
represented by the new data frame}
}
\value{
A data frame representing one sensitivity score from the
Barbour et al. 1996 method
}
\description{
Barbour Sensitivity Score
}
