% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{bde}
\alias{bde}
\title{BDE}
\usage{
bde(prep.sensitivity, metric.col)
}
\arguments{
\item{prep.sensitivity}{PARAM_DESCRIPTION}

\item{metric.col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
 \code{\link[purrr]{map}}
}
