% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_funs.R
\name{check_dups}
\alias{check_dups}
\title{Check for Duplicates}
\usage{
check_dups(x, ...)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{...}{A column or set of columns to aggreate by and check for duplicates.}
}
\value{
A data frame containing the counts of the variables in the specified
column(s). Values greater than one indicate duplicates.
}
\description{
Check the data frame for duplicates based on the specified
column(s).
}
