% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_funs.R
\name{clean_df}
\alias{clean_df}
\title{Clean Up Data Frames}
\usage{
clean_df(x)
}
\arguments{
\item{x}{A data frame or tibble.}
}
\value{
A data frame.
}
\description{
Performs some data frame clean up that I have found to
reduce errors and make it easier to work with the data frame. All
of the column names are converted to all lower case and any leading
or trailing white space is removed. Similarly, all character columns
are converted to all lower case and any leading or trailing white space
is removed.
}
