% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_funs.R
\name{fill_taxa}
\alias{fill_taxa}
\title{Fill NAs with Previous Taxonomic Rank}
\usage{
fill_taxa(long.df, final.id.col, ...)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{final.id.col}{The name of the column that contains the final taxonomic
ID.}

\item{...}{A set of columns related to taxonomic ranks where NA
values will be replaced with the previous taxonomic rank. It is important to
list these columns in the appropriate order (e.g. phylum to species). The
order the columns are presented is used to create the hierarchy that
identifies the previous taxonomic rank.}
}
\value{
A data frame.
}
\description{
Fill in NA values present in the taxonomic hierarchy with the
previous taxonomic rank.
}
