% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_plot.R
\name{index_plot2}
\alias{index_plot2}
\title{Index Plot}
\usage{
index_plot2(index.df, condition.colname, condition.order = NULL,
  index.colname = "INDEX", threshold, plot.title, rating.df)
}
\arguments{
\item{scores.df}{= data frame of metric values for each station}

\item{ref.cond}{= The site classification that represents better
environmental conditions.}

\item{deg.cond}{= The site classification that represents the degraded
environmental conditions.}

\item{method}{= the sensitivity function to be used during the assessment.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Index Plot
}
