% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_rating.R
\name{index_rating}
\alias{index_rating}
\title{Index Rating}
\usage{
index_rating(
  scores.df,
  index.colname,
  condition.colname = NULL,
  ref.cond = NULL
)
}
\arguments{
\item{scores.df}{= a long data frame format containing the index scores.}

\item{index.colname}{= the name of the column containing the index scores.}

\item{condition.colname}{= the name of the column containing the disturbance
gradient classes.}

\item{ref.cond}{= the character string that refers to the Reference class in
the disturbance gradient column (condition.colname).}
}
\value{
A wrapper function that allows you to select your scoring methodology
and a metric sensitivity threshold.
}
\description{
Index Rating
}
