% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_score.R
\name{index_score}
\alias{index_score}
\title{Index Score}
\usage{
index_score(scores.df, first.metric, keep.metrics)
}
\arguments{
\item{scores.df}{= data frame of metric values for each station}

\item{ref.cond}{= The site classification that represents better
environmental conditions.}

\item{deg.cond}{= The site classification that represents the degraded
environmental conditions.}

\item{method}{= the sensitivity function to be used during the assessment.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Index Score
}
