% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_plots.R
\name{metric_plots}
\alias{metric_plots}
\title{Metric Plots}
\usage{
metric_plots(metrics.df, condition.colname, first.metric, order.condition,
  include.title = TRUE, name.pdf)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}
}
\value{
Exports a pdf of metric box-and-whisker plots into your working 
directory.
}
\description{
Metric Plots
}
