% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{nonpar_tbl}
\alias{nonpar_tbl}
\title{Nonparametric Test Table}
\usage{
nonpar_tbl(wide.df, first.metric, group.by)
}
\arguments{
\item{wide.df}{= a data frame of samples to be scored.}

\item{first.metric}{= the first metric column that appears in your wide
data frame when reading from left to right. It is assumed that all columns
to the right of the first metric column are metrics. Any other columns will
result in an error or, if the column is numeric, it will be treated as if
it were a metric.}

\item{groub.by}{= specify the column upon which the nonparametric assessment
will be performed. If the group.by column contains less than 2 groups, then
an error will be returned. If the group.by column contains 2 groups, then
a Wilcoxon Rank Sum Test will be performed. If the group.by column contains
more than 2 columns, then a Kruskal-Wallis Test and a post hoc Dunn's test
will be performed.}
}
\value{
Nonparametric test results are summarized into a clean table.
}
\description{
Nonparametric Test Table
}
