% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orginal_de.R
\name{original_de}
\alias{original_de}
\title{Discrimination Efficiency}
\usage{
original_de(deg.df, quant.df, first.metric)
}
\arguments{
\item{deg.df}{= a data frame of only the lower.class values.}

\item{quant.df}{= Data frame containing upper.class quantile values.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Discrimination Efficiency
}
