% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_helpers.R
\name{prep_taxa_df}
\alias{prep_taxa_df}
\title{Return the Data Frame to the Order Orginally Provided}
\usage{
prep_taxa_df(.data, .key_col, .unnest_col, .filter)
}
\arguments{
\item{.data}{The data frame}

\item{.key_col}{The name of the column that contains a unique sampling.}

\item{.unnest_col}{Name the nested column}

\item{.filter}{supply a logical statement}
}
\value{
A data frame.
}
\description{
Return the data frame to the order orginally provided.
}
