% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{prob_rarefaction}
\alias{prob_rarefaction}
\title{Probabilistic Rarefaction}
\usage{
prob_rarefaction(long.df, unique.id.col, count.col, taxon.col, sample)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}

\item{sample}{The size of the subsample.}
}
\value{
A data frame.
}
\description{
Standardized subsample.
}
