% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_ref_gradient.R
\name{ref_gradient}
\alias{ref_gradient}
\title{Scoring Method: Reference Gradient}
\usage{
ref_gradient(metrics.df, sensitivity.df, first.metric, condition.colname,
  ref.cond)
}
\arguments{
\item{long.df}{<- a data frame of samples to be scored.}
}
\value{
Scores the samples.
}
\description{
Scoring Method: Reference Gradient
}
