% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_metrics.R
\name{score_metrics}
\alias{score_metrics}
\title{Score Metrics}
\usage{
score_metrics(metrics.df, scoring.method, sensitivity.df,
  sensitivity.colname, sensitivity.threshold = 0, first.metric,
  condition.colname = NULL, ref.cond = NULL)
}
\arguments{
\item{metrics.df}{<- a data frame of samples to be scored.}
}
\value{
A wrapper function that allows you to select your scoring methodology
and a metric sensitivity threshold.
}
\description{
Score Metrics
}
