% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{select_thresh}
\alias{select_thresh}
\title{Select A Sensitivity Threshold}
\usage{
select_thresh(x, thresh.col)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{thresh.col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
 \code{\link[purrr]{map}}
}
