% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_de.R
\name{sensitivity_de}
\alias{sensitivity_de}
\title{Sensitivity: Discrimination Efficiency (DE)}
\usage{
sensitivity_de(metrics.df, condition.col, ref.cond, deg.cond, ...)
}
\arguments{
\item{metrics.df}{PARAM_DESCRIPTION}

\item{condition.col}{PARAM_DESCRIPTION}

\item{ref.cond}{PARAM_DESCRIPTION}

\item{deg.cond}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarize}},\code{\link[dplyr]{ungroup}},\code{\link[dplyr]{select}},\code{\link[dplyr]{full_join}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}},\code{\link[dplyr]{left_join}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}}
 \code{\link[rlang]{quos}},\code{\link[rlang]{enquo}}
 \code{\link[tidyr]{gather}}
}
