% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_youden.R
\name{sensitivity_youden}
\alias{sensitivity_youden}
\title{Sensitivity: Youden's Index}
\usage{
sensitivity_youden(metrics.df, condition.col, ref.cond, deg.cond, ...)
}
\arguments{
\item{metrics.df}{PARAM_DESCRIPTION}

\item{condition.col}{PARAM_DESCRIPTION}

\item{ref.cond}{PARAM_DESCRIPTION}

\item{deg.cond}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarize}},\code{\link[dplyr]{ungroup}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{case_when}},\code{\link[dplyr]{quo_name}},\code{\link[dplyr]{bind_rows}},\code{\link[dplyr]{left_join}},\code{\link[dplyr]{}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{select}},\code{\link[dplyr]{arrange}},\code{\link[dplyr]{desc}}
 \code{\link[OptimalCutpoints]{optimal.cutpoints}},\code{\link[OptimalCutpoints]{control.cutpoints}}
 \code{\link[rlang]{quos}},\code{\link[rlang]{enquo}},\code{\link[rlang]{sym}}
 \code{\link[tidyr]{gather}},\code{\link[tidyr]{spread}}
}
