% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_sensitivity.R
\name{seq_sensitivity}
\alias{seq_sensitivity}
\title{Chunk Sensitivity}
\usage{
seq_sensitivity(metrics.df, condition.colname, ref.cond = "REF",
  deg.cond = "DEG", method)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{method}{= the sensitivity function to be used during the assessment.}

\item{upper.class}{= the site class that represents the better condition.}

\item{lower.class}{= the site class that represents the poorer condition.}
}
\value{
Determines the threshold at which a metric best categorizes
two defined environmental conditions.
}
\description{
Chunk Sensitivity
}
