% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{spread_condition}
\alias{spread_condition}
\title{Spread Condition}
\usage{
spread_condition(
  .dataframe,
  .metric_col,
  .condition_col,
  .value_col,
  .reference,
  .degraded
)
}
\arguments{
\item{.dataframe}{PARAM_DESCRIPTION}

\item{.metric_col}{PARAM_DESCRIPTION}

\item{.condition_col}{PARAM_DESCRIPTION}

\item{.value_col}{PARAM_DESCRIPTION}

\item{.reference}{PARAM_DESCRIPTION}

\item{.degraded}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}},\code{\link[rlang]{sym}},\code{\link[rlang]{quo_name}}
}
