% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_div.R
\name{taxa_div}
\alias{taxa_div}
\title{Percentage of a Specified Taxon}
\usage{
taxa_div(
  .data,
  .key_col,
  .counts_col,
  .group_col,
  .filter = NULL,
  .job,
  .base_log,
  .q,
  .unnest_col = data
)
}
\arguments{
\item{.data}{Taxonomic counts arranged in a long data format.}

\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{.counts_col}{The name of the column that contains taxanomic counts.}

\item{.group_col}{The name of the column that contains the taxon or taxa
of interest.}

\item{.keep_vec}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
