% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_pct.R
\name{taxa_pct}
\alias{taxa_pct}
\title{Percentage of a Specified .keep_vec}
\usage{
taxa_pct(.data, .key_col, .counts_col, .filter = NULL, .unnest_col = data)
}
\arguments{
\item{.data}{.keep_vecomic counts arranged in a long data format.}

\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{.counts_col}{The name of the column that contains taxanomic counts.}

\item{.keep_col}{The name of the column that contains the taxon or taxa
of interest.}

\item{.keep_vec}{The .keep_vec or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified .keep_vec or taxa.
}
