% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_pct_rich.R
\name{taxa_pct_rich}
\alias{taxa_pct_rich}
\title{Percentage of a Specified Taxon}
\usage{
taxa_pct_rich(long.df, unique.id.col, count.col, taxon.col, taxon = NULL,
  high.res.taxa.col, exclusion.col = NULL, exclusion.vec = NULL)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}

\item{high.res.taxa.col}{The name of the column that contains the taxon or taxa
of interest.}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
