% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_rich.R
\name{taxa_rich}
\alias{taxa_rich}
\title{Taxon Richness}
\usage{
taxa_rich(long.df, low.taxa.col, high.taxa.col = NULL, taxon = NULL,
  exclusion.col = NULL, exclusion.vec = NULL, na.rm = TRUE)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}
}
\value{
The number of taxa identified.
}
\description{
Taxon Richness
}
