% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_tol_index.R
\name{taxa_tol_index}
\alias{taxa_tol_index}
\title{Percentage of a Specified Taxon}
\usage{
taxa_tol_index(
  .data,
  .key_col,
  .counts_col,
  .group_col,
  .filter,
  .tol_col,
  na.rm = TRUE
)
}
\arguments{
\item{.data}{Taxonomic counts arranged in a long data format.}

\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{.counts_col}{The name of the column that contains taxanomic counts.}

\item{.group_col}{The name of the column that contains the taxon or taxa
of interest.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
