% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{test_thresh}
\alias{test_thresh}
\title{Test Sensitivity Thresholds}
\usage{
test_thresh(x, less.than.col, greater.than.col, thresh)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{less.than.col}{PARAM_DESCRIPTION}

\item{greater.than.col}{PARAM_DESCRIPTION}

\item{thresh}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
}
