% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barbour.R
\name{barbour}
\alias{barbour}
\title{Barbour et al. 1996 method for testing metric sensitivity}
\usage{
barbour()
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}

\item{ref.df}{= a data frame of only the upper.class values.}

\item{deg.df}{= a data frame of only the lower.class values.}
}
\value{
A data frame of metrics scored using the Barbour et al. 1996 method
for testing metric sensitivity
}
\description{
Barbour et al. 1996 method for testing metric sensitivity
}
