% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{de_orginal}
\alias{de_orginal}
\title{Discrimination Efficiency}
\usage{
de_orginal(prep.sensitivity, metric.col)
}
\arguments{
\item{prep.sensitivity}{PARAM_DESCRIPTION}

\item{metric.col}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}}
}
