% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orginal_order.R
\name{original_order}
\alias{original_order}
\title{Return the Data Frame to the Order Orginally Provided}
\usage{
original_order(mod.df, original.df, unique.id.col)
}
\arguments{
\item{mod.df}{The data frame altered during metric calculation.}

\item{original.df}{The orginal data frame, and therefore the original order of the unique.id.col.}

\item{unique.id.col}{The name of the column that contains a unique sampling.}
}
\value{
A data frame.
}
\description{
Return the data frame to the order orginally provided.
}
