% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_helpers.R
\name{original_order}
\alias{original_order}
\title{Return the Data Frame to the Order Orginally Provided}
\usage{
original_order(.mod_data, .org_data, .key_col)
}
\description{
Return the Data Frame to the Order Orginally Provided
}
