% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{prep_sensitivity}
\alias{prep_sensitivity}
\title{Prepare Sensitivity}
\usage{
prep_sensitivity(metrics.long, metric.col, value.col, condition.col,
  ref.cond, deg.cond)
}
\arguments{
\item{metrics.long}{PARAM_DESCRIPTION}

\item{metric.col}{PARAM_DESCRIPTION}

\item{value.col}{PARAM_DESCRIPTION}

\item{condition.col}{PARAM_DESCRIPTION}

\item{ref.cond}{PARAM_DESCRIPTION}

\item{deg.cond}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{enquo}},\code{\link[rlang]{quo_name}}
}
