% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Metric Sensitivity}
\usage{
sensitivity(metrics.long, first.metric, condition.colname, ref.cond = "REF",
  deg.cond = "DEG", method = "BDE")
}
\arguments{
\item{ref.cond}{= The site classification that represents better
environmental conditions.}

\item{deg.cond}{= The site classification that represents the degraded
environmental conditions.}

\item{method}{= the sensitivity function to be used during the assessment.}

\item{metrics.df}{= data frame of metric values for each station}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Metric Sensitivity
}
