% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_abund.R
\name{taxa_abund}
\alias{taxa_abund}
\title{Abundance of a Specified Taxon}
\usage{
taxa_abund(
  .dataframe,
  .key_col,
  .counts_col,
  .filter = NULL,
  .unnest_col = data
)
}
\arguments{
\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{.counts_col}{The name of the column that contains taxanomic counts.}

\item{.data}{Taxonomic counts arranged in a long data format.}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the abundance of each sample represented by the
specified taxon or taxa.
}
