% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_abund.R
\name{taxa_abund}
\alias{taxa_abund}
\title{Abundance of a Specified Taxon}
\usage{
taxa_abund(long.df, count.col, taxon.col, taxon = NULL,
  exclusion.col = NULL, exclusion.vec = NULL)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the abundance of each sample represented by the
specified taxon or taxa.
}
