% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_div.R
\name{taxa_div}
\alias{taxa_div}
\title{Taxonomic Diversity}
\usage{
taxa_div(
  .dataframe,
  .key_col,
  .counts_col,
  .group_col,
  .filter = NULL,
  .job,
  .base_log,
  .q,
  .unnest_col = NULL
)
}
\arguments{
\item{.dataframe}{A data frame where each row should represent the number of
individuals enumerated for a single taxon collected during a single sampling event.}

\item{.key_col}{One unquoted column name that represents a key (i.e., unique ID)
for a sampling event for which to group (i.e., aggregate) the data.}

\item{.counts_col}{One unquoted column name that represents taxonomic counts.}

\item{.group_col}{One unquoted column name that represents a taxomic rank
or group of interest.}

\item{.filter}{A logical statement to subset the data frame prior to calculating
the metric of interest.}

\item{.unnest_col}{One unqouted column name that represents nested data.
If this column is NULL (default), then the data will not be unnested.}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
