% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_div.R
\name{taxa_div}
\alias{taxa_div}
\title{Percentage of a Specified Taxon}
\usage{
taxa_div(long.df, unique.id.col, count.col, low.taxa.col, high.taxa.col,
  taxon = NULL, job, base.log, q)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{high.taxa.col}{The name of the column that contains the taxon or taxa
of interest.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
