% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{taxa_pct}
\alias{taxa_pct}
\title{Percentage of a Specified Taxon}
\usage{
taxa_pct(long.df, unique.id.col, count.col, taxon.col, taxon, keep.na = FALSE)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
A numeric vector of percentages.
}
\description{
Calculate the percentage of each sample represented by the
specified taxon or taxa.
}
