% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_rich.R
\name{taxa_rich}
\alias{taxa_rich}
\title{Taxon Richness}
\usage{
taxa_rich(.data, .key_col, .group_col, .filter = NULL, .unnest_col = data)
}
\arguments{
\item{.data}{Taxonomic counts arranged in a long data format.}

\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}

\item{.keep_vec}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}
}
\value{
The number of taxa identified.
}
\description{
Taxon Richness
}
