% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness.R
\name{taxa_rich}
\alias{taxa_rich}
\title{Taxon Richness}
\usage{
taxa_rich(long.df, unique.id.col, low.taxa.col, high.taxa.col, taxon = NULL,
  count.na = TRUE)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{taxon}{The taxon or taxa of interest. To specify more than one taxa
use: c("TAXA1", "TAXA2", "TAXA3").}

\item{taxon.col}{The name of the column that contains the taxon or taxa
of interest.}
}
\value{
The number of taxa identified.
}
\description{
Taxon Richness
}
