% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_seq.R
\name{taxa_seq}
\alias{taxa_seq}
\title{The percentage each taxon makes up of a sample}
\usage{
taxa_seq(
  .data,
  .key_col,
  .counts_col,
  .filter_cols_vec,
  .group_col,
  .unnest_col = data,
  .keep_na = FALSE,
  .job,
  .base_log = NULL,
  .q = NULL
)
}
\arguments{
\item{.data}{Taxonomic counts arranged in a long data format.}

\item{.key_col}{The name of the column that contains a unique sampling
event ID.}

\item{.counts_col}{The name of the column that contains taxanomic counts.}

\item{.job}{To calculate the percent of each taxon specify "pct". To calculate
the richness of each tax}

\item{.filter_cols_vecs}{The name of the column(s) that contains the taxa
of interest.}
}
\value{
The percent of the sample represented by each taxa per taxon level.
}
\description{
The percentage each taxon makes up of a sample
}
