% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{taxa_seq}
\alias{taxa_seq}
\title{The percentage each taxon makes up of a sample}
\usage{
taxa_seq(long.df, unique.id.col, count.col, taxa.cols, high.taxa.col = NULL,
  keep.na = FALSE, job)
}
\arguments{
\item{long.df}{Taxonomic counts arranged in a long data format.}

\item{unique.id.col}{The name of the column that contains a unique sampling
event ID.}

\item{count.col}{The name of the column that contains taxanomic counts.}

\item{job}{To calculate the percent of each taxon specify "pct". To calculate
the richness of each tax}

\item{taxon.cols}{The name of the column(s) that contains the taxa
of interest.}
}
\value{
The percent of the sample represented by each taxa per taxon level.
}
\description{
The percentage each taxon makes up of a sample
}
