% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_seq.R
\name{taxa_seq}
\alias{taxa_seq}
\title{Sequence Through Taxonomic Metrics}
\usage{
taxa_seq(
  .dataframe,
  .key_col,
  .counts_col,
  .filter_cols_vec,
  .group_col = NULL,
  .unnest_col = NULL,
  .job,
  .base_log = 2,
  .q = NULL
)
}
\arguments{
\item{.dataframe}{A data frame where each row should represent the number of
individuals enumerated for a single taxon collected during a single sampling event.}

\item{.key_col}{One unquoted column name that represents a key (i.e., unique ID)
for a sampling event for which to group (i.e., aggregate) the data.}

\item{.counts_col}{One unquoted column name that represents taxonomic counts.}

\item{.filter_cols_vec}{A quoted vector of column names for which
the function will sequence through each unique value to perform the specified .job.}

\item{.group_col}{One unquoted column name that represents a taxomic rank
or group of interest.}

\item{.unnest_col}{One unqouted column name that represents nested data.
If this column is NULL (default), then the data will not be unnested.}

\item{.job}{A character string specifying the metric of interest.
Below is a list of exceptable inputs:
\itemize{
 \item{"abund"}{Description needed}
 \item{"pct"}{Description needed}
 \item{"rich"}{Description needed}
 \item{"pct_rich"}{Description needed}
 \item{"shannon"}{Description needed}
 \item{"effective_shannon"}{Description needed}
 \item{"simpson""}{Description needed}
 \item{"invsimpson"}{Description needed}
 \item{"gini_simpson"}{Description needed}
 \item{"effective_simpson"}{Description needed}
 \item{"pielou"}{Description needed}
 \item{"margalef"}{Description needed}
 \item{"menhinick"}{Description needed}
 \item{"hill"}{Description needed}
 \item{"renyi"}{Description needed}
}}

\item{.base_log}{The base log value used during the calculation of
Shannon Diversity index ("shannon") or Effective Shannon Diversity ("effective_shannon").
The default value is two.}

\item{.q}{The exponent used during the calculation of Hill Numbers ("hill") and
Renyi Entropy ("renyi").}
}
\value{
A data frame where each column represents numeric metric values.
}
\description{
This function is a wrapper for the other "taxa_" functions that sequences
through unique values within specified columns. The intent is to provide the ability to
quickly calculate metrics for a large number of taxa without having to call a "taxa_"
function for each taxon of interest.
}
